+++ src/onegov.swissvotes/onegov/swissvotes/app.py
@@ -133,6 +133,7 @@ def get_frameworks_asset():
     yield 'jquery.datetimepicker.css'
     yield 'jquery.datetimepicker.js'
     yield 'datetimepicker.js'
+    yield 'dropzone.js'
 
 
 @SwissvotesApp.webasset('common')