+++ src/onegov.swissvotes/onegov/swissvotes/layouts/__init__.py
@@ -1,8 +1,10 @@
 from onegov.swissvotes.layouts.default import DefaultLayout
 from onegov.swissvotes.layouts.mail import MailLayout
 from onegov.swissvotes.layouts.page import AddPageLayout
+from onegov.swissvotes.layouts.page import DeletePageAttachmentLayout
 from onegov.swissvotes.layouts.page import DeletePageLayout
 from onegov.swissvotes.layouts.page import EditPageLayout
+from onegov.swissvotes.layouts.page import PageAttachmentsLayout
 from onegov.swissvotes.layouts.page import PageLayout
 from onegov.swissvotes.layouts.vote import DeleteVoteLayout
 from onegov.swissvotes.layouts.vote import UploadVoteAttachemtsLayout
@@ -16,11 +18,13 @@ from onegov.swissvotes.layouts.votes import VotesLayout
 __all__ = (
     'AddPageLayout',
     'DefaultLayout',
+    'DeletePageAttachmentLayout',
     'DeletePageLayout',
     'DeleteVoteLayout',
     'DeleteVotesLayout',
     'EditPageLayout',
     'MailLayout',
+    'PageAttachmentsLayout',
     'PageLayout',
     'UpdateVotesLayout',
     'UploadVoteAttachemtsLayout',