+++ src/onegov.swissvotes/onegov/swissvotes/theme/swissvotes_theme.py
@@ -19,6 +19,7 @@ class SwissvotesTheme(BaseTheme):
             'form',
             'table',
             'alert',
+            'dropzone',
             'swissvotes'
         ]